+++ HISTORY.md
@@ -3,3 +3,8 @@
 
 ### 03/28/2012
 * Version bump: changes power button from 'select' to 'back' (change made to device directory).
+
+### 03/30/2012:
+* Move reboot menu into advanced menu and change normal menu reboot option back to 'reboot system now'.
+* Changed fastboot option to say 'fastboot bootloader' instead of just 'fastboot mode' to hopefully negate confusion.
+* Version bump.